src/requirements.txt
Flask==1.1.2
google-cloud-storage
google-cloud-logging
google-auth
google-api-python-client
google_auth_oauthlib
google-cloud-tasks
googleapis-common-protos
requests
redis
